requirements.txt
cffi==1.15.1
click==8.1.3
colorama==0.4.4
crayons==0.4.0
cryptography==38.0.4
Flask==2.1.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
numpy==1.22.4
opencv-python==4.5.5.64
piexif==1.1.3
Pillow==9.1.1
protobuf==3.20.1
pycparser==2.21
pygame==2.1.2
stegano==0.10.2
Werkzeug==2.1.2
